Revert "Use lfe-1.1.1 Hex package"
This reverts commit 78eccf2bbda3b1b536fb9529a7bab5b1ee9496dc.

The change of moving to Hex was not approved. This was discussed in detail on
slack, complete with a pinned message about the reasons. I will be developing a
different approach that supports hex, github, and gitlab, per the preference of
the developer -- without forcing any developer to use any one mirror, but
rather giving them the freedom of their preferences.